requirements.txt
cudatoolkit==10.0.130
editdistance==0.5.3
fairseq==0.9.0
jieba==0.42.1
numpy==1.19.0
matplotlib==3.2.1
opencv-python==4.3.0.36
Pillow==7.1.2
numpy==1.19.0
lmdb==0.98
PyYAML==5.3.1
six==1.14.0
tqdm==4.45.0
transformers==2.8.0
torch==1.1.0
torchvision==0.3.0
tqdm==4.32.2
python==3.7.7
